src/requirements.txt
PyPDF2
googlemaps
python-dotenv
colorama